+++ ormar/fields/base.py
@@ -1,3 +1,4 @@
+import inspect
 from typing import Any, List, Optional, TYPE_CHECKING, Type, Union
 
 import sqlalchemy